# Optimize nth and nth_back for BoundTupleIterator