--- a/newsfragments/4897.added.md
+++ b/newsfragments/4897.added.md
@@ -1,0 +1,1 @@
+Optimizes `nth`, `nth_back`, `advance_by` and `advance_back_by` for `BoundTupleIterator`